+++ Cargo.nix
@@ -5108,7 +5108,7 @@ rec {
         workspace_member = null;
         src = pkgs.fetchgit {
           url = "https://github.com/stackabletech//operator-rs.git";
-          rev = "8b1b821d6913d31a589cf7db5d089d982b220fb1";
+          rev = "01535bc1acb05e8a28f098d40bf5bef3105e075e";
           sha256 = "0p43kywyfjq9l3j0fapzpdcdphy8nyn7l9n4i9isd1iwpyn5ix3f";
         };
         libName = "k8s_version";
@@ -9995,7 +9995,7 @@ rec {
         workspace_member = null;
         src = pkgs.fetchgit {
           url = "https://github.com/stackabletech//operator-rs.git";
-          rev = "8b1b821d6913d31a589cf7db5d089d982b220fb1";
+          rev = "01535bc1acb05e8a28f098d40bf5bef3105e075e";
           sha256 = "0p43kywyfjq9l3j0fapzpdcdphy8nyn7l9n4i9isd1iwpyn5ix3f";
         };
         libName = "stackable_certs";
@@ -10098,7 +10098,7 @@ rec {
         workspace_member = null;
         src = pkgs.fetchgit {
           url = "https://github.com/stackabletech//operator-rs.git";
-          rev = "8b1b821d6913d31a589cf7db5d089d982b220fb1";
+          rev = "01535bc1acb05e8a28f098d40bf5bef3105e075e";
           sha256 = "0p43kywyfjq9l3j0fapzpdcdphy8nyn7l9n4i9isd1iwpyn5ix3f";
         };
         libName = "stackable_operator";
@@ -10297,7 +10297,7 @@ rec {
         workspace_member = null;
         src = pkgs.fetchgit {
           url = "https://github.com/stackabletech//operator-rs.git";
-          rev = "8b1b821d6913d31a589cf7db5d089d982b220fb1";
+          rev = "01535bc1acb05e8a28f098d40bf5bef3105e075e";
           sha256 = "0p43kywyfjq9l3j0fapzpdcdphy8nyn7l9n4i9isd1iwpyn5ix3f";
         };
         procMacro = true;
@@ -10332,7 +10332,7 @@ rec {
         workspace_member = null;
         src = pkgs.fetchgit {
           url = "https://github.com/stackabletech//operator-rs.git";
-          rev = "8b1b821d6913d31a589cf7db5d089d982b220fb1";
+          rev = "01535bc1acb05e8a28f098d40bf5bef3105e075e";
           sha256 = "0p43kywyfjq9l3j0fapzpdcdphy8nyn7l9n4i9isd1iwpyn5ix3f";
         };
         libName = "stackable_shared";
@@ -10413,7 +10413,7 @@ rec {
         workspace_member = null;
         src = pkgs.fetchgit {
           url = "https://github.com/stackabletech//operator-rs.git";
-          rev = "8b1b821d6913d31a589cf7db5d089d982b220fb1";
+          rev = "01535bc1acb05e8a28f098d40bf5bef3105e075e";
           sha256 = "0p43kywyfjq9l3j0fapzpdcdphy8nyn7l9n4i9isd1iwpyn5ix3f";
         };
         libName = "stackable_telemetry";
@@ -10523,7 +10523,7 @@ rec {
         workspace_member = null;
         src = pkgs.fetchgit {
           url = "https://github.com/stackabletech//operator-rs.git";
-          rev = "8b1b821d6913d31a589cf7db5d089d982b220fb1";
+          rev = "01535bc1acb05e8a28f098d40bf5bef3105e075e";
           sha256 = "0p43kywyfjq9l3j0fapzpdcdphy8nyn7l9n4i9isd1iwpyn5ix3f";
         };
         libName = "stackable_versioned";
@@ -10573,7 +10573,7 @@ rec {
         workspace_member = null;
         src = pkgs.fetchgit {
           url = "https://github.com/stackabletech//operator-rs.git";
-          rev = "8b1b821d6913d31a589cf7db5d089d982b220fb1";
+          rev = "01535bc1acb05e8a28f098d40bf5bef3105e075e";
           sha256 = "0p43kywyfjq9l3j0fapzpdcdphy8nyn7l9n4i9isd1iwpyn5ix3f";
         };
         procMacro = true;
@@ -10641,7 +10641,7 @@ rec {
         workspace_member = null;
         src = pkgs.fetchgit {
           url = "https://github.com/stackabletech//operator-rs.git";
-          rev = "8b1b821d6913d31a589cf7db5d089d982b220fb1";
+          rev = "01535bc1acb05e8a28f098d40bf5bef3105e075e";
           sha256 = "0p43kywyfjq9l3j0fapzpdcdphy8nyn7l9n4i9isd1iwpyn5ix3f";
         };
         libName = "stackable_webhook";